# Speedup Add::__hash__()