--- a/symengine/add.cpp
+++ b/symengine/add.cpp
@@ -50,12 +50,12 @@ bool Add::is_canonical(const RCP<const Number> &coef,
 
 std::size_t Add::__hash__() const
 {
-    std::size_t seed = ADD;
+    std::size_t seed = ADD, temp;
     hash_combine<Basic>(seed, *coef_);
-    map_basic_num ordered(dict_.begin(), dict_.end());
-    for (const auto &p: ordered) {
-        hash_combine<Basic>(seed, *(p.first));
-        hash_combine<Basic>(seed, *(p.second));
+    for (const auto &p: dict_) {
+        temp = p.first->hash();
+        hash_combine<Basic>(temp, *(p.second));
+        seed += temp;
     }
     return seed;
 }