I checked benchmarks. Improvements in `expand6` (from 5604ms to 5077ms) and `expand7` (4586ms to 3752ms). No visible change in `expand2` and `add1`