@certik 2015-12-29
+1 to merge if tests pass.