@certik 2015-12-29
@isuruf see also #719, I can't see much speedup, but maybe you can see some speedup in some benchmarks.